add carts

+++ models/product.go
@@ -9,14 +9,14 @@ import (
 
 type Products struct {
 	gorm.Model
-	Title    string  `gorm:"size:255;not null"`
-	Details  string  `gorm:"not null"`
-	SortDesc string  `gorm:"not null" text:"html"`
-	Slug     string  `gorm:"size:255;unique_index;not null"`
-	Price    float32 `gorm:"not null"`
-	Quantity int     `gorm:"not null"`
-	CatID    int     `gorm:"not null"`
-	SubcatID int     `gorm:"not null"`
+	Title    string `gorm:"size:255;not null"`
+	Details  string `gorm:"not null"`
+	SortDesc string `gorm:"not null" text:"html"`
+	Slug     string `gorm:"size:255;unique_index;not null"`
+	Price    uint32 `gorm:"not null"`
+	Quantity int    `gorm:"not null"`
+	CatID    int    `gorm:"not null"`
+	SubcatID int    `gorm:"not null"`
 	// Categories []Category `gorm:"many2many:products_categories;"`
 
 	Images string `gorm:"not null"`
@@ -34,7 +34,7 @@ type ProductList struct {
 	CatID    int
 	SubcatID int
 	Details  template.HTML
-	Price    float32
+	Price    float64
 	Quantity int
 	Images   string
 }
